requirements.txt
altair==4.2.2
altair_saver==0.5.0
fpdf2==2.8.2
matplotlib==3.10.7
numpy==2.3.4
pandas==2.3.3
Pillow==12.0.0
plotly==5.22.0
scipy==1.16.2
streamlit==1.42.1
streamlit_option_menu==0.3.13
tzlocal==5.2
pyJWT==2.8.0

